Refactoring of index page
